admin section added

+++ app/Http/Controllers/CategoriesController.php
@@ -93,6 +93,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count()>0){
+            session()->flash('error','Category cannot be Deleted because it has some posts');
+            return redirect()->back();
+        }
         $category->delete();
         session()->flash('success','Category Deleted Successfully');
         return redirect(route('categories.index'));
