+++ app/Http/Controllers/TagsController.php
@@ -93,6 +93,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count()>0){
+            session()->flash('error','Tag cannot be Deleted because it has some posts');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success','Tags Deleted Successfully');
         return redirect(route('tags.index'));
